# docs/swagger.yaml
basePath: /api
definitions:
...
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  CondidatDetails:
    description: CondidatDetails represents detailed information about a specific
      condidat.
    properties:
      companyID:
        description: CompanyID is the unique identifier for the company associated
          with the condidat.
        type: string
      companyName:
        description: CompanyName is the name of the company associated with the condidat.
        type: string
      educationLevel:
        type: string
      firstname:
        description: Name is the name of the condidat.
        type: string
      id:
        description: ID is the unique identifier for the condidat.
        type: string
      lastname:
        description: last name of condidat
        type: string
      university:
        type: string
    type: object
  CondidatIn:
    description: CondidatIn represents the input structure for creating a new condidat.
    properties:
      adress:
        maxLength: 35
        minLength: 3
        type: string
      companyID:
        description: CompanyID is the unique identifier for the company associated
          with the user. It is required.
        type: string
      education_level:
        maxLength: 35
        minLength: 3
        type: string
      email:
        description: Email is the email address of the user. It is required, should
          be a valid email, and maximum length is 255 characters.
        maxLength: 255
        type: string
      firstName:
        description: Firstname is the first name of the user. It is required and should
          be between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
      lastName:
        description: Lastname is the last name of the user. It is required and should
          be between 3 and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      password:
        description: Password is the user's password. It is required, and its length
          should be between 10 and 255 characters.
        maxLength: 255
        minLength: 10
        type: string
      university:
        maxLength: 35
        minLength: 3
        type: string
    required:
    - adress
    - companyID
    - education_level
    - email
    - firstName
    - lastName
    - password
    - university
    type: object
  CondidatsCount:
    description: CondidatsCount represents the count of condidats.
    properties:
      count:
        description: Count is the number of condidats.
        type: integer
    type: object
...
  CondidtasPagination:
    description: CondidtasPagination represents the paginated list of Condidats.
    properties:
      items:
        description: Items is a slice containing individual condidat details.
        items:
          $ref: '#/definitions/CondidatsTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of condidats in the entire list.
        type: integer
    type: object
...
  LoggedIn:
    description: LoggedIn represents the candidat details after successful login.
    properties:
      ID:
        description: ID is the unique identifier for the candidat.
        type: string
      email:
        description: Email is the email address of the candidat.
        type: string
      first_name:
        description: Name is the name of the candidat.
        type: string
      last_name:
        description: Name is the name of the candidat.
        type: string
      workCompanyId:
        description: CompanyID is the unique identifier for the candidat company.
        type: string
    type: object
...
  Signin:
    description: Signin represents the information required for signing in candidat.
    properties:
      email:
        description: Email is the email address of the user. It is required, should
          be a valid email, and maximum length is 255 characters.
        maxLength: 255
        type: string
      password:
        description: Password is the user's password. It is required, and its length
          should be between 10 and 255 characters.
        maxLength: 255
        minLength: 10
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /condidats/{companyID}:
    get:
      description: Get all Candidats .
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CondidtasPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Candidats
      tags:
      - Condidats
    post:
      consumes:
      - application/json
      description: Create a new candidate.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: condidat query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CondidatIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create candidate
      tags:
      - Condidats
  /condidats/{companyID}/{ID}:
    delete:
      description: "Delete one candidat\t."
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: candidat ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete candidat
      tags:
      - Condidats
    get:
      description: Get one candidat.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: candidat ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CondidatDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get candidat
      tags:
      - Condidats
    put:
      consumes:
      - application/json
      description: Update one candidat.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: candidat ID
        in: path
        name: ID
        required: true
        type: string
      - description: candidat query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CondidatIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update candidat
      tags:
      - Condidats
  /condidats/{companyID}/Signin:
    post:
      consumes:
      - application/json
      description: Authenticate and log in a uscandidater.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: candidat query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Signin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoggedIn'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      summary: Signin
      tags:
      - Condidats
  /condidats/{companyID}/count:
    get:
      description: Get number of all Candidats .
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CondidatsCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get number of  Candidats
      tags:
      - Condidats
  /condidats/{companyID}/list:
    get:
      description: Get list of all Candidats.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/CondidatsList'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list of  Candidats
      tags:
      - Condidats
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"